src/requirements.txt
cookiecutter-data-science
python-dotenv
beautifulsoup4
pySmartDL
psycopg2
pandas
Jinja2
typer
rarfile

# YOUTUBE API
simple-youtube-api
google-auth-oauthlib
google-api-python-client

# loguru